Fixing a compiler warning.

+++ src/map_shp.c
@@ -165,7 +165,7 @@ void create_shapefile_map(char *dir, char *shapefile_name, int type,
             shapefile_name);
 
         // Change spaces to underlines
-        for (ii = 0; ii < strlen(temp_shapefile_name); ii++) {
+        for (ii = 0; ii < (int)strlen(temp_shapefile_name); ii++) {
             if (temp_shapefile_name[ii] == ' ') {
                 temp_shapefile_name[ii] = '_';
             }
